v10.17: ENGLISH WORDS! Double-precision RMS fixes logit distribution
Changed all RMSNorm sum-of-squares from float to double precision,
matching llama.cpp's ggml_float=double convention. This sharpened
the logit distribution from peak=8.9 to peak=12.9, enabling
multi-character tokens to dominate over single characters.

Generated output: "Cancel..:V..ext....Cancel dismiss."
- First real ENGLISH WORDS from the custom engine!
- Multi-character tokens: Cancel, ext, dismiss
- Not yet coherent (not thinking about "Hello") but REAL WORDS

Changed in: attention.h rmsnorm(), deltanet_impl.h dn_rmsnorm_weighted(),
deltanet_impl.h dn_rmsnorm_simple().

The 397B model is now producing English text from a fully custom
C+CUDA engine. 11 bugs fixed + double-precision RMS = working output.

+++ engine/runtime/deltanet_impl.h
@@ -81,19 +81,19 @@ static void dn_l2norm(float* x, int n) {
 
 /* RMS norm (no learned weight version) */
 static void dn_rmsnorm_simple(float* out, const float* x, int n) {
-    float ss = 0.0f;
+    double ss = 0.0;
     int i;
-    for (i = 0; i < n; i++) ss += x[i] * x[i];
-    float inv = 1.0f / sqrtf(ss / n + 1e-6f);
+    for (i = 0; i < n; i++) ss += (double)x[i] * (double)x[i];
+    float inv = 1.0f / sqrtf((float)(ss / n) + 1e-6f);
     for (i = 0; i < n; i++) out[i] = x[i] * inv;
 }
 
 /* RMS norm with learned weight */
 static void dn_rmsnorm_weighted(float* out, const float* x, const float* w, int n) {
-    float ss = 0.0f;
+    double ss = 0.0;
     int i;
-    for (i = 0; i < n; i++) ss += x[i] * x[i];
-    float inv = 1.0f / sqrtf(ss / n + 1e-6f);
+    for (i = 0; i < n; i++) ss += (double)x[i] * (double)x[i];
+    float inv = 1.0f / sqrtf((float)(ss / n) + 1e-6f);
     for (i = 0; i < n; i++) out[i] = x[i] * inv * w[i];
 }
